``` [Feat] Implement MyScans feature to display animal scan data and refactor API integration
This commit introduces the MyScans feature, allowing users to view details about scanned animals and their associated images. It also refactors the API integration to support fetching animal tracks data.

- Added `MyScansScreenSuccess` to display detailed information about animal scans, including the species name and number of scans.
- Implemented image loading in `MyScansScreenSuccess` using Coil for asynchronous image retrieval.
- Refactored `AnimalDataModel` to include `images`, `nomFr`, `nomLatin`, `populationEstimee`, and more fields.
- Created `AnimalTracksApiService` and `AnimalTracksRepository` for fetching animal tracks data from the API.
- Implemented `AnimalTracksRepositoryImpl` to handle API requests for animal tracks.
- Updated `Repository` to include `AnimalTracksRepository` and `getAnimalTracks()` method.
- Updated `MyScansState` to include `Success` state with `imagesTracks` property.
- Modified `MyScansViewModel` to fetch and manage the animal tracks data.
- Updated `MyScansViewModel` to display a success message.
- Updated `MyScansViewModel` to load animal tracks data.
- Updated `MyScansScreen` to display animal tracks data.
- Updated `DataModule` to provide `AnimalTracksApiService` and `AnimalTracksRepository`.
- Updated `MetaDataModel` to include `populationEstimee` field as Double.
- Added Coil dependency to `build.gradle.kts`.
```

+++ app/src/main/java/com/wildlens/mspr_2025/data/models/MetaDataModel.kt
@@ -22,5 +22,5 @@ data class MetaDataModel(
     @SerialName("nom_latin")
     val nomLatin: String,
     @SerialName("population_estimee")
-    val populationEstimee: Long
+    val populationEstimee: Double
 )
